% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_faceted_hist_plot.R
\name{create_faceted_hist_plot}
\alias{create_faceted_hist_plot}
\title{Make faceted plot}
\usage{
create_faceted_hist_plot(df, feature)
}
\arguments{
\item{df}{Tidy data containing a list of genres and acoustic features, as a dataframe}

\item{feature}{Specified acoustic feature of which we want a visualisation, as a character type}
}
\value{
faceted plot of 6 histograms
}
\description{
Creates a faceted plot for the "playlist_genre" target variable
including any selected features
combined using the face_grid() function
}
\examples{
x <- RCurl::getURL("https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2020/2020-01-21/spotify_songs.csv")
test_data <- read.csv(text = x)
create_faceted_hist_plot(test_data, 'danceability')
}
